backend/requirements.txt
pymupdf
python-docx
python-multipart
scikit-learn>=1.0.0
numpy>=1.20.0